Fixed bugs and updated the tests
